Florida_Beach_Insider/requirements/dev-requirements.txt
######### POLITIFACT Development dependencies ###############

django==1.11.11
psycopg2-binary
django-extensions
django-debug-toolbar
python-decouple
